Add IAS capability to loadable_basic application
Added BLE capability to the capsense_loadable_basic application,
including the Immediate Alert Service (IAS). This allows the user to
connect to the device remotely and switch to the DFU app (app0). Before
this addition, the user would have to manually reset the device while
holding down SW2 in order to enter the bootloader.